reorganized plotting functions
