+++ tests/benchmarks/conftest.py
@@ -313,21 +313,260 @@ def sandbox_runoff(sandbox_qext: torch.Tensor) -> torch.Tensor:
 
 
 # =============================================================================
+# DDR Mock Classes
+# =============================================================================
+
+# Reach IDs in the Sandbox network (RAPID2 ordering)
+RAPID2_REACH_IDS = [10, 20, 30, 40, 50]
+
+
+class MockStreamflow(torch.nn.Module):
+    """Mock streamflow module that returns pre-interpolated Q' (lateral inflow).
+
+    In real DDR, the streamflow module predicts Q' from forcing data.
+    For testing, we directly use interpolated Qext as the Q'.
+    """
+
+    def __init__(self, qprime: torch.Tensor) -> None:
+        """Initialize with pre-computed Q'.
+
+        Parameters
+        ----------
+        qprime : torch.Tensor
+            Interpolated hourly lateral inflow, shape (timesteps, reaches)
+        """
+        super().__init__()
+        self.qprime = qprime
+
+    def forward(self, **_kwargs) -> torch.Tensor:
+        """Return the pre-computed Q'."""
+        return self.qprime
+
+
+class MockKAN(torch.nn.Module):
+    """Mock KAN module that returns fixed normalized parameters.
+
+    All parameters are set to 0.5 (normalized midpoint), which maps to
+    the geometric mean for log-space parameters and linear mean otherwise.
+    """
+
+    def __init__(self, num_reaches: int, learnable_params: list[str]) -> None:
+        """Initialize with fixed parameter values.
+
+        Parameters
+        ----------
+        num_reaches : int
+            Number of river reaches
+        learnable_params : list[str]
+            List of parameter names to output (e.g., ["n", "q_spatial", "top_width", "side_slope"])
+        """
+        super().__init__()
+        self.num_reaches = num_reaches
+        self.learnable_params = learnable_params
+
+    def forward(self, **_kwargs) -> dict[str, torch.Tensor]:
+        """Return fixed normalized parameters (all 0.5)."""
+        return {
+            param: torch.full((self.num_reaches,), 0.5, dtype=torch.float32)
+            for param in self.learnable_params
+        }
+
+
+# =============================================================================
+# DDR Helper Functions
+# =============================================================================
+
+
+def create_ddr_config():
+    """Create a minimal Config for DDR routing tests.
+
+    Returns
+    -------
+    ddr.validation.configs.Config
+        Pydantic config with default DDR parameter ranges
+    """
+    from omegaconf import DictConfig
+
+    from ddr.validation.configs import validate_config
+
+    cfg_dict = {
+        "name": "sandbox_test",
+        "mode": "testing",
+        "geodataset": "merit",
+        "data_sources": {
+            "geospatial_fabric_gpkg": "mock.gpkg",
+            "streamflow": "mock://streamflow",
+            "conus_adjacency": "mock.zarr",
+        },
+        "params": {
+            "parameter_ranges": {
+                "n": [0.015, 0.25],  # Manning's n (log space)
+                "q_spatial": [0.0, 1.0],  # Shape factor (linear)
+                "top_width": [1.0, 5000.0],  # Channel width (log space)
+                "side_slope": [0.5, 50.0],  # Side slope (log space)
+            },
+            "log_space_parameters": ["n", "top_width", "side_slope"],
+            "defaults": {"p_spatial": 21},
+            "attribute_minimums": {
+                "velocity": 0.01,
+                "depth": 0.01,
+                "discharge": 0.0001,
+                "bottom_width": 0.1,
+                "slope": 0.0001,
+            },
+            "tau": 3,
+        },
+        "kan": {
+            "input_var_names": ["mock"],
+            "learnable_parameters": ["n", "q_spatial", "top_width", "side_slope"],
+        },
+        "s3_region": "us-east-2",
+        "device": "cpu",
+    }
+    return validate_config(DictConfig(cfg_dict), save_config=False)
+
+
+def create_routing_dataclass(sandbox_zarr_path: Path, num_reaches: int = 5):
+    """Create a RoutingDataclass for DDR routing with sandbox data.
+
+    Parameters
+    ----------
+    sandbox_zarr_path : Path
+        Path to zarr store containing adjacency matrix
+    num_reaches : int
+        Number of river reaches
+
+    Returns
+    -------
+    tuple[RoutingDataclass, list[int]]
+        Routing dataclass configured for sandbox network and the topological order
+    """
+    from ddr_engine.merit import coo_from_zarr
+
+    from ddr.geodatazoo.dataclasses import Dates, RoutingDataclass
+
+    # Load adjacency matrix from zarr and convert to torch sparse CSR
+    coo, ts_order = coo_from_zarr(sandbox_zarr_path)
+    coo_csr = coo.tocsr()
+
+    # Convert scipy CSR to torch sparse CSR tensor
+    crow_indices = torch.from_numpy(coo_csr.indptr).to(torch.int32)
+    col_indices = torch.from_numpy(coo_csr.indices).to(torch.int32)
+    values = torch.from_numpy(coo_csr.data).to(torch.float32)
+    adjacency_matrix = torch.sparse_csr_tensor(
+        crow_indices, col_indices, values, size=coo_csr.shape, dtype=torch.float32
+    )
+
+    # Create Dates object for the sandbox period
+    dates = Dates(start_time="1970/01/01", end_time="1970/01/10")
+
+    routing_dataclass = RoutingDataclass(
+        adjacency_matrix=adjacency_matrix,
+        length=torch.full((num_reaches,), 5000.0),  # 5km reaches
+        slope=torch.full((num_reaches,), 0.001),  # 1m per km
+        x=torch.full((num_reaches,), 0.25),  # Storage coefficient from x_Sandbox.csv
+        top_width=torch.empty(0),  # Learned via spatial_params
+        side_slope=torch.empty(0),  # Learned via spatial_params
+        divide_ids=np.array(ts_order),
+        outflow_idx=None,  # Output all segments
+        dates=dates,
+        observations=None,
+        gage_catchment=None,
+    )
+
+    return routing_dataclass, ts_order
+
+
+def reorder_ddr_to_rapid2(ddr_output: np.ndarray, ts_order: list[int]) -> np.ndarray:
+    """Reorder DDR output from topological order to RAPID2 order [10, 20, 30, 40, 50].
+
+    Parameters
+    ----------
+    ddr_output : np.ndarray
+        DDR discharge output in topological order, shape (reaches, timesteps)
+    ts_order : list[int]
+        Topological order of reach IDs from zarr store
+
+    Returns
+    -------
+    np.ndarray
+        Discharge reordered to RAPID2 order [10, 20, 30, 40, 50]
+    """
+    # Build mapping: for each RAPID2 position, find the index in ts_order
+    reorder_idx = [ts_order.index(rid) for rid in RAPID2_REACH_IDS]
+    return ddr_output[reorder_idx, :]
+
+
+def run_ddr_routing(sandbox_zarr_path: Path, sandbox_hourly_qprime: torch.Tensor) -> np.ndarray:
+    """Run DDR routing and return discharge output in RAPID2 order.
+
+    Parameters
+    ----------
+    sandbox_zarr_path : Path
+        Path to zarr store containing adjacency matrix
+    sandbox_hourly_qprime : torch.Tensor
+        Interpolated hourly Q', shape (timesteps, reaches) in RAPID2 order
+
+    Returns
+    -------
+    np.ndarray
+        Discharge output in RAPID2 order [10, 20, 30, 40, 50], shape (5, timesteps)
+    """
+    from ddr import dmc
+
+    num_reaches = 5
+    learnable_params = ["n", "q_spatial", "top_width", "side_slope"]
+
+    # Create components
+    cfg = create_ddr_config()
+    routing_dataclass, ts_order = create_routing_dataclass(sandbox_zarr_path, num_reaches)
+
+    # Reorder Q' from RAPID2 order to topological order for DDR input
+    qprime_rapid2 = sandbox_hourly_qprime.numpy()  # (timesteps, 5) in RAPID2 order
+    topo_idx = [RAPID2_REACH_IDS.index(rid) for rid in ts_order]
+    qprime_topo = qprime_rapid2[:, topo_idx]  # Reorder to topo order
+    qprime_tensor = torch.from_numpy(qprime_topo).float()
+
+    mock_streamflow = MockStreamflow(qprime_tensor)
+    mock_kan = MockKAN(num_reaches, learnable_params)
+    routing_model = dmc(cfg=cfg, device="cpu")
+
+    # Get Q' from mock streamflow
+    qprime = mock_streamflow()
+
+    # Get spatial params from mock KAN
+    spatial_params = mock_kan()
+
+    # Run DDR routing
+    routing_model.set_progress_info(epoch=0, mini_batch=0)
+    dmc_kwargs = {
+        "routing_dataclass": routing_dataclass,
+        "spatial_parameters": spatial_params,
+        "streamflow": qprime,
+    }
+    ddr_output = routing_model(**dmc_kwargs)
+    ddr_discharge_topo = ddr_output["runoff"].detach().numpy()  # (reaches, timesteps) in topo order
+
+    # Reorder output from topological order to RAPID2 order
+    ddr_discharge_rapid2 = reorder_ddr_to_rapid2(ddr_discharge_topo, ts_order)
+
+    return ddr_discharge_rapid2
+
+
+# =============================================================================
 # DDR Routing Fixtures
 # =============================================================================
 
 
 @pytest.fixture(scope="session")
 def ddr_discharge(sandbox_zarr_path: Path, sandbox_hourly_qprime: torch.Tensor) -> np.ndarray:
-    """Run DDR routing and return discharge output.
+    """Run DDR routing and return discharge output in RAPID2 order.
 
     This fixture provides DDR discharge for use in comparison plots.
 
     Returns
     -------
     np.ndarray
-        Discharge output, shape (5, 238) for (reaches, timesteps)
+        Discharge output in RAPID2 order [10, 20, 30, 40, 50], shape (5, 238)
     """
-    from tests.benchmarks.test_ddr import run_ddr_routing
-
     return run_ddr_routing(sandbox_zarr_path, sandbox_hourly_qprime)
